+++ indexer/data_header.cpp
@@ -33,7 +33,7 @@ namespace feature
 
   void DataHeader::SetScales(int * arr)
   {
-    for (int i = 0; i < m_scales.size(); ++i)
+    for (size_t i = 0; i < m_scales.size(); ++i)
       m_scales[i] = static_cast<uint8_t>(arr[i]);
   }
